Refactor class OuTreeSynchronisation.
