+++ src/main/java/de/muenchen/zammad/ldap/SyncApplication.java
@@ -6,7 +6,7 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.context.ConfigurableApplicationContext;
 
 import de.muenchen.zammad.ldap.config.RoleIds;
-import de.muenchen.zammad.ldap.sync.OuTreeControl;
+import de.muenchen.zammad.ldap.sync.TreeControl;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -20,7 +20,7 @@ public class SyncApplication {
 
 		RoleIds complementRoleIds = context.getBean(RoleIds.class);
 		if (complementRoleIds.isAddAllRoleIdsSuccessful()) {
-		    OuTreeControl syncService = context.getBean(OuTreeControl.class);
+		    TreeControl syncService = context.getBean(TreeControl.class);
 			syncService.synchronizationControl();;
 		}
 		else
