+++ src/test/java/de/muenchen/zammad/ldap/service/CreateGroupAndUserTest.java
@@ -20,7 +20,9 @@ import org.mockito.quality.Strictness;
 
 import de.muenchen.zammad.domain.Group;
 import de.muenchen.zammad.domain.User;
-import de.muenchen.zammad.ldap.sync.OuTreeSynchronization;
+import de.muenchen.zammad.ldap.sync.EmailAddressCache;
+import de.muenchen.zammad.ldap.sync.TreeSynchronization;
+import de.muenchen.zammad.ldap.sync.SignatureCache;
 import de.muenchen.zammad.ldap.sync.ZammadService;
 import de.muenchen.zammad.ldap.tree.LdapOuNode;
 import lombok.extern.log4j.Log4j2;
@@ -52,7 +54,7 @@ class CreateGroupAndUserTest extends PrepareTestEnvironment {
         userAndGroupMocks(zammadService);
         channelsMock(zammadService);
 
-		var zammadSyncServiceSubtree = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+		var zammadSyncServiceSubtree = new TreeSynchronization(zammadService, createZammadProperties(), new EmailAddressCache(zammadService, standardDefaultMock()), new SignatureCache(zammadService, standardDefaultMock()));
 
 		zammadSyncServiceSubtree.updateZammadGroupsWithUsers(createLdapTree());
 
@@ -83,7 +85,7 @@ class CreateGroupAndUserTest extends PrepareTestEnvironment {
         groupMocksCreateParentNodeTest(zammadService);
         channelsMock(zammadService);
 
-        var zammadSyncService = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+        var zammadSyncService = new TreeSynchronization(zammadService, createZammadProperties(), new EmailAddressCache(zammadService, standardDefaultMock()), new SignatureCache(zammadService, standardDefaultMock()));
 
         var childTree_level_2 = new TreeMap<String, LdapOuNode>();
         var number = 1;
