+++ src/test/java/de/muenchen/zammad/ldap/service/CreateUpdateGroupAndUserTest.java
@@ -19,7 +19,9 @@ import org.mockito.quality.Strictness;
 
 import de.muenchen.zammad.domain.Group;
 import de.muenchen.zammad.domain.User;
-import de.muenchen.zammad.ldap.sync.OuTreeSynchronization;
+import de.muenchen.zammad.ldap.sync.EmailAddressCache;
+import de.muenchen.zammad.ldap.sync.TreeSynchronization;
+import de.muenchen.zammad.ldap.sync.SignatureCache;
 import de.muenchen.zammad.ldap.sync.ZammadService;
 
 @ExtendWith(MockitoExtension.class)
@@ -51,7 +53,7 @@ class CreateUpdateGroupAndUserTest extends PrepareTestEnvironment {
         userAndGroupMocks(zammadService);
         channelsMock(zammadService);
 
-		var zammadSyncServiceSubtree = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+		var zammadSyncServiceSubtree = new TreeSynchronization(zammadService, createZammadProperties(), new EmailAddressCache(zammadService, standardDefaultMock()), new SignatureCache(zammadService, standardDefaultMock()));
 
 		var ldapTree = createLdapTree();
 
@@ -96,7 +98,7 @@ class CreateUpdateGroupAndUserTest extends PrepareTestEnvironment {
         userAndGroupMocks(zammadService);
         channelsMock(zammadService);
 
-        var zammadSyncService = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+        var zammadSyncService = new TreeSynchronization(zammadService, createZammadProperties(), new EmailAddressCache(zammadService, standardDefaultMock()), new SignatureCache(zammadService, standardDefaultMock()));
 
         var ldapTree = createLdapTree();
