+++ src/test/java/de/muenchen/zammad/ldap/service/DeleteGroupAndUserTest.java
@@ -19,8 +19,8 @@ import org.mockito.quality.Strictness;
 
 import de.muenchen.zammad.domain.Group;
 import de.muenchen.zammad.domain.User;
-import de.muenchen.zammad.ldap.sync.OuTreeControl;
-import de.muenchen.zammad.ldap.sync.OuTreeSynchronization;
+import de.muenchen.zammad.ldap.sync.DeletedLdapUser;
+import de.muenchen.zammad.ldap.sync.TreeControl;
 import de.muenchen.zammad.ldap.sync.ZammadService;
 
 
@@ -55,15 +55,15 @@ class DeleteGroupAndUserTest extends PrepareTestEnvironment {
         assertEquals(1, zammadService.getZammadGroups().size());
         assertEquals(21, zammadService.getZammadUsers().size());
 
-		var zammadSyncServiceSubtree = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+        var deletedLdapUser = new DeletedLdapUser(zammadService);
 
 		var reducedLdapTree = reducedLdapTree();
 		var rootNode = reducedLdapTree.entrySet().iterator().next().getValue();
 		assertEquals(17, rootNode.flatListLdapUserDTO().size());
 
-		var reducedEnhancedLdapUserDTO = OuTreeControl.allLdapUsersWithDistinguishedNames(reducedLdapTree);
+		var reducedEnhancedLdapUserDTO = TreeControl.collectUserFromAllBranches(reducedLdapTree);
 
-		zammadSyncServiceSubtree.assignDeletionFlagZammadUser(rootNode, reducedEnhancedLdapUserDTO);
+		deletedLdapUser.checkForRemoval(rootNode, reducedEnhancedLdapUserDTO);
 
 		verify(zammadService, times(1)).updateZammadUser(updateUserCaptor.capture());
 		assertEquals("delete", updateUserCaptor.getAllValues().get(0).getLdapsyncstate());
@@ -83,15 +83,15 @@ class DeleteGroupAndUserTest extends PrepareTestEnvironment {
         assertEquals(7, zammadService.getZammadGroups().size());
         assertEquals(21, zammadService.getZammadUsers().size());
 
-        var zammadSyncServiceSubtree = new OuTreeSynchronization(zammadService, createZammadProperties(), standardDefaultMock());
+        var deletedLdapUser = new DeletedLdapUser(zammadService);
 
         var reducedLdapTree = reducedLdapTree();
         var rootNode = reducedLdapTree.entrySet().iterator().next().getValue();
         assertEquals(17, rootNode.flatListLdapUserDTO().size());
 
-        var reducedEnhancedLdapUserDTO = OuTreeControl.allLdapUsersWithDistinguishedNames(reducedLdapTree);
+        var reducedEnhancedLdapUserDTO = TreeControl.collectUserFromAllBranches(reducedLdapTree);
 
-        zammadSyncServiceSubtree.assignDeletionFlagZammadUser(rootNode, reducedEnhancedLdapUserDTO);
+        deletedLdapUser.checkForRemoval(rootNode, reducedEnhancedLdapUserDTO);
 
         verify(zammadService, times(4)).updateZammadUser(updateUserCaptor.capture());
         assertEquals("delete", updateUserCaptor.getAllValues().get(0).getLdapsyncstate());
